Update the lease returned when claiming an offer to include names
In order to allow the lease to include the names, we have to move
lease/offer controller functions into utils.
